requirements.txt
Django==3.1.3
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==4.4.0
djangorestframework-files==1.1.0
mysqlclient==2.0.1
Pillow==8.0.1
django-upload-validator==1.1.5
requests


